requirements.txt
appdirs==1.4.4
asgiref==3.5.2
certifi==2022.9.14
charset-normalizer==2.1.1
Django==4.1.1
django-crispy-forms==1.14.0
django-environ==0.9.0
idna==3.4
lxml==4.9.1
multitasking==0.0.11
numpy==1.23.3
pandas==1.5.0
python-coinmarketcap==0.3
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
yfinance==0.1.77
